refactor: review notes fix
