+++ src/components/Popup.js
@@ -1,15 +1,15 @@
 export default class Popup {
-  constructor(popupSelector) {
-    this._popupSelector = popupSelector;
+  constructor(popup) {
+    this._popup = popup;
   }
 
   open() {
-    this._popupSelector.classList.add('popup_opened');
+    this._popup.classList.add('popup_opened');
     document.addEventListener('keydown', this._handleEscClose);
   }
 
   close() {
-    this._popupSelector.classList.remove('popup_opened');
+    this._popup.classList.remove('popup_opened');
     document.removeEventListener('keydown', this._handleEscClose);
   }
 
@@ -20,8 +20,8 @@ export default class Popup {
   }
 
   setEventListeners() {
-    this._popupSelector.addEventListener('click', (evt) => {
-      if (evt.target.classList.contains('popup') || evt.target.classList.contains('popup__close-button')) {
+    this._popup.addEventListener('mousedown', (evt) => {
+      if (evt.target.classList.contains('popup__overlay') || evt.target.classList.contains('popup__close-button')) {
         this.close();
       };
     })
