+++ src/components/Section.js
@@ -1,11 +1,11 @@
 export default class Section {
-  constructor({ renderItems }, containerSelector) {
+  constructor({ renderItems }, container) {
     this._renderer = renderItems;
-    this._containerSelector = containerSelector;
+    this._container = container;
   }
 
   addItem(element) {
-    this._containerSelector.prepend(element);
+    this._container.prepend(element);
   }
 
   renderItems(items) {
